refactor: separate browser entities and data access
